pyproject.toml
[project]
name = "magic-coder"
version = "0.1.0"
description = "Magic Coder - AI-powered code assistant family"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "wxd123", email = "wxd123@users.noreply.github.com"},
]
license = "MIT"
dependencies = [
    "magicc-shared",
    "openai>=2.33.0",
    "psutil>=7.2.2",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


# 关键：配置空包
[tool.hatch.build.targets.wheel]
packages = [
    "./packages/comment/src/magicc_comment",
    "./packages/code/src/magicc_code", 
    "./packages/arch/src/magicc_arch",
    "./packages/test/src/magicc_test",
    "./packages/report/src/magicc_report"
]

[tool.uv.workspace]
members = [
    "packages/*",
    "shared",
]

[tool.uv.sources]
magicc-comment = { path = "packages/comment", editable = true }
magicc-code = { path = "packages/code", editable = true }
magicc-arch = { path = "packages/arch", editable = true }
magicc-test = { path = "packages/test", editable = true }
magicc-report = { path = "packages/report", editable = true }
magicc-shared = { workspace = true, editable = true }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests", "packages/*/tests", "shared/tests"]
addopts = "-v --cov=magicc --cov-report=html"

[dependency-groups]
dev = [
    "build>=1.4.4",
    "twine>=6.2.0",
]
